refactor: 스터디 공지사항 테이블을 공유 UI 패키지로 이동 #112
- 관리자 및 인트라 앱의 스터디 공지사항 테이블 컴포넌트를 공유 UI 패키지의 StudyAnnouncementTable로 통합 이동
- 새로운 공유 컴포넌트 사용을 위한 관리자 및 인트라 탭 섹션 업데이트로 코드 재사용성 및 유지보수성 개선
- 공지사항용 데스크톱 및 모바일 컬럼 정의 추가 및 export 업데이트
