+++ packages/ui/src/components/study/StudyAnnouncementTable/announcement-columns.tsx
@@ -1,0 +1,3 @@
+// Re-export desktop and mobile column definitions
+export { STUDY_ANNOUNCEMENT_COLUMN } from './desktop-announcement-columns';
+export { MOBILE_STUDY_ANNOUNCEMENT_COLUMN } from './mobile-announcement-columns';
