+++ packages/ui/src/components/study/index.ts
@@ -1,0 +1,1 @@
+export { StudyAnnouncementTable } from './StudyAnnouncementTable/StudyAnnouncementTable';
