+++ packages/ui/src/index.ts
@@ -27,6 +27,7 @@ export * from './components/label/anchorLabel';
 export * from './components/label/title';
 export * from './components/layout/content-section';
 export * from './components/layout/page-layout';
+export * from './components/study';
 export * from './components/study/lecture-list-item';
 export * from './components/study/study-unassigned-group';
 export * from './components/study/add-group-dialog/add-group-dialog';
@@ -57,6 +58,7 @@ export * from './components/study/study-info-section';
 export * from './components/study/mobile-study-button';
 export * from './components/study/study-group-list-item';
 export * from './components/study/study-group-list';
+export * from './components/study/StudyAnnouncementTable/StudyAnnouncementTable';
 export * from './components/select/checkbox-list';
 export * from './components/select/labeled-checkbox-list';
 export * from './constants/assets';
